Moved advanced options to inline (instead of a separate modal) and moved to consistent filenames
