+++ ui/src/main/webapp/src/app/app.module.ts
@@ -9,22 +9,22 @@ import {FileSelectDirective, FileDropDirective, FileUploader} from 'ng2-file-upl
 import { AppComponent }  from './components/app.component';
 import { routing, appRoutingProviders } from './app.routing';
 
-import {ProjectListComponent} from "./components/projectlist.component";
-import {AnalysisContextFormComponent} from "./components/analysiscontextform.component";
-import {ApplicationGroupForm} from "./components/applicationgroupform.component";
-import {GroupListComponent} from "./components/grouplist.component";
-import {MigrationProjectFormComponent} from "./components/migrationprojectform.component";
+import {ProjectListComponent} from "./components/project-list.component";
+import {AnalysisContextFormComponent} from "./components/analysis-context-form.component";
+import {ApplicationGroupForm} from "./components/application-group-form.component";
+import {GroupListComponent} from "./components/group-list.component";
+import {MigrationProjectFormComponent} from "./components/migration-project-form.component";
 import {RegisterApplicationFormComponent} from "./components/registerapplicationform.component";
-import {ProgressBarComponent} from "./components/progressbar.component";
+import {ProgressBarComponent} from "./components/progress-bar.component";
 import {NavbarComponent} from "./components/navbar.component";
 import {BreadCrumbsComponent} from "./components/breadcrumbs.component";
 import {ConfigurationService} from "./services/configuration.service";
-import {AnalysisContextService} from "./services/analysiscontext.service";
-import {ApplicationGroupService} from "./services/applicationgroup.service";
+import {AnalysisContextService} from "./services/analysis-context.service";
+import {ApplicationGroupService} from "./services/application-group.service";
 import {FileService} from "./services/file.service";
-import {MigrationPathService} from "./services/migrationpath.service";
-import {MigrationProjectService} from "./services/migrationproject.service";
-import {RegisteredApplicationService} from "./services/registeredapplication.service";
+import {MigrationPathService} from "./services/migration-path.service";
+import {MigrationProjectService} from "./services/migration-project.service";
+import {RegisteredApplicationService} from "./services/registered-application.service";
 import {WindupService} from "./services/windup.service";
 import {RuleService} from "./services/rule.service";
 import {ConfigurationComponent} from "./components/configuration.component";
