+++ ui/src/main/webapp/src/app/components/analysis-context-advanced-options-modal.component.ts
@@ -8,64 +8,51 @@ import {ValidationResult} from "../model/validation-result.model";
 @Component({
     selector: 'analysis-context-advanced-options',
     template: `
-    <modal-dialog>
-        <div header>
-            Advanced Options
-        </div>
-        <div body>
-            <form class="form-horizontal">
-                
-                <table class="datatable table table-striped table-bordered">
-                    <thead>
-                        <tr>
-                            <th width="33%">Option</th>
-                            <th width="50%">Value</th>
-                            <th width="17%">Actions</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <tr *ngFor="let selectedOption of selectedOptions; let i = index;">
-                            <td>
-                                {{selectedOption.name}}
-                            </td>
-                            <td>
-                                {{selectedOption.value}}
-                            </td>
-                            <td>
-                                <button (click)="removeAdvancedOption(i)" class="btn-warning" href="#">Delete</button>
-                            </td>
-                        </tr>
-                        <tr *ngIf="newOption">
-                            <td class="input-group">
-                                <select class="form-control" name="newOptionTypeSelection" [(ngModel)]="newOption.name" (change)="newOptionTypeChanged()">
-                                    <option *ngFor="let option of availableOptions" value="{{option.name}}">{{option.name}}</option>
+            <table class="datatable table table-striped table-bordered">
+                <thead>
+                    <tr>
+                        <th width="33%">Option</th>
+                        <th width="50%">Value</th>
+                        <th width="17%">Actions</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    <tr *ngFor="let selectedOption of selectedOptions; let i = index;">
+                        <td>
+                            {{selectedOption.name}}
+                        </td>
+                        <td>
+                            {{selectedOption.value}}
+                        </td>
+                        <td>
+                            <button (click)="removeAdvancedOption(i)" class="btn-warning" href="#">Delete</button>
+                        </td>
+                    </tr>
+                    <tr *ngIf="newOption">
+                        <td class="input-group">
+                            <select class="form-control" name="newOptionTypeSelection" [(ngModel)]="newOption.name" (change)="newOptionTypeChanged()">
+                                <option *ngFor="let option of availableOptions" value="{{option.name}}">{{option.name}}</option>
+                            </select>
+                            <w-popover class="input-group-addon" *ngIf="currentSelectedOptionDefinition?.description" [content]="currentSelectedOptionDefinition?.description"></w-popover>
+                        </td>
+                        <td [class.bg-danger]="newOptionError" align="right">
+                            <span class="text-danger">{{newOptionError}}</span>
+                            <div [ngSwitch]="currentOptionType">
+                                <input *ngSwitchCase="'text'" type="text" name="currentOptionInput" class="form-control" [(ngModel)]="newOption.value">
+                                <input *ngSwitchCase="'checkbox'" type="checkbox" class="form-control" name="currentOptionInput" [(ngModel)]="newOption.value">
+                                <select *ngSwitchCase="'select'" class="form-control" name="newOptionSelect" [(ngModel)]="newOption.value">
+                                    <option *ngFor="let option of currentSelectedOptionDefinition.availableValues" value="{{option}}">{{option}}</option>
                                 </select>
-                                <w-popover class="input-group-addon" *ngIf="currentSelectedOptionDefinition?.description" [content]="currentSelectedOptionDefinition?.description"></w-popover>
-                            </td>
-                            <td [class.bg-danger]="newOptionError" align="right">
-                                <span class="text-danger">{{newOptionError}}</span>
-                                <div [ngSwitch]="currentOptionType">
-                                    <input *ngSwitchCase="'text'" type="text" name="currentOptionInput" class="form-control" [(ngModel)]="newOption.value">
-                                    <input *ngSwitchCase="'checkbox'" type="checkbox" class="form-control" name="currentOptionInput" [(ngModel)]="newOption.value">
-                                    <select *ngSwitchCase="'select'" class="form-control" name="newOptionSelect" [(ngModel)]="newOption.value">
-                                        <option *ngFor="let option of currentSelectedOptionDefinition.availableValues" value="{{option}}">{{option}}</option>
-                                    </select>
-                                </div>
-                            </td>
-                            <td>
-                                <button (click)="addAdvancedOption()">Add</button>
-                            </td>
-                        </tr>
-                    </tbody>
-                </table>
-                <button *ngIf="!newOption" (click)="startAddNew()">Add New Option</button>
-                
-            </form>
-        </div>
-        <div footer>
-            <button type="button" class="btn btn-primary" (click)="hide()">Done</button>
-        </div>
-    </modal-dialog>
+                            </div>
+                        </td>
+                        <td>
+                            <button (click)="addAdvancedOption($event)" class="btn btn-default">Add</button>
+                            <button (click)="cancelAddAdvancedOption($event)" class="btn btn-default">Cancel</button>
+                        </td>
+                    </tr>
+                </tbody>
+            </table>
+            <button *ngIf="!newOption" (click)="startAddNew()" class="btn btn-default">Add New Option</button>
 `
 })
 export class AnalysisContextAdvancedOptionsModalComponent {
@@ -77,9 +64,6 @@ export class AnalysisContextAdvancedOptionsModalComponent {
 
     @Output()
     advancedOptionsChanged:EventEmitter<AdvancedOption[]> = new EventEmitter<AdvancedOption[]>();
-
-    @ViewChild(ModalDialogComponent)
-    private modalDialog:ModalDialogComponent;
 
     private newOption:AdvancedOption;
     private newOptionError:string;
@@ -141,9 +125,16 @@ export class AnalysisContextAdvancedOptionsModalComponent {
     private startAddNew() {
         this.newOption = <AdvancedOption>{};
         this.newOption.value = "";
+        return false;
     }
 
-    private addAdvancedOption() {
+    private cancelAddAdvancedOption() {
+        this.newOption = null;
+        return false;
+    }
+
+    private addAdvancedOption(event:Event) {
+        event.preventDefault();
         this.newOptionError = "";
         // Only accept null for a checkbox (with a checkbox "null" == false).
         if (this.currentOptionType != 'checkbox' && (this.newOption.value == null || this.newOption.value == "")) {
@@ -178,13 +169,5 @@ export class AnalysisContextAdvancedOptionsModalComponent {
     newOptionTypeChanged() {
         if (this.newOption)
             this.newOption.value = "";
-    }
-
-    show() {
-        this.modalDialog.show();
-    }
-
-    hide() {
-        this.modalDialog.hide();
     }
 }
