+++ ui/src/main/webapp/src/app/components/analysis-context-form.component.html
@@ -52,8 +52,13 @@
         <custom-rule-selection [selectedRulePaths]="analysisContext.rulesPaths" (selectedRulePathsChanged)="rulesPathsChanged($event)"></custom-rule-selection>
         <div class="form-group">
             <label class="col-md-2 control-label">Advanced Options</label>
-            <div class="col-md-3">
-                <button class="btn btn-default" (click)="viewAdvancedOptions(advancedOptionsModal)">View Advanced Options</button>
+            <div class="col-md-6">
+                <analysis-context-advanced-options
+                        [(selectedOptions)]="analysisContext.advancedOptions"
+                        (advancedOptionsChanged)="advancedOptionsChanged($event)"
+                        [configurationOptions]="configurationOptions"
+                        #advancedOptionsModal>
+                </analysis-context-advanced-options>
             </div>
         </div>
         <div class="form-group">
@@ -66,4 +71,3 @@
     </div>
 </form>
 
-<analysis-context-advanced-options [(selectedOptions)]="analysisContext.advancedOptions" (advancedOptionsChanged)="advancedOptionsChanged($event)" [configurationOptions]="configurationOptions" #advancedOptionsModal></analysis-context-advanced-options>
