+++ ui/src/main/webapp/src/app/components/edit-application-form.component.ts
@@ -3,16 +3,16 @@ import {FormBuilder, Validators} from "@angular/forms";
 import {ActivatedRoute, Router} from "@angular/router";
 import {FileUploader} from "ng2-file-upload/ng2-file-upload";
 import {RegisteredApplication} from "../windup-services";
-import {RegisteredApplicationService} from "../services/registeredapplication.service";
+import {RegisteredApplicationService} from "../services/registered-application.service";
 import {FileExistsValidator} from "../validators/file-exists.validator";
 import {FileService} from "../services/file.service";
-import {ApplicationGroupService} from "../services/applicationgroup.service";
+import {ApplicationGroupService} from "../services/application-group.service";
 import {Constants} from "../constants";
 import {KeycloakService} from "../services/keycloak.service";
 import {RegisterApplicationFormComponent} from "./registerapplicationform.component";
 
 @Component({
-    templateUrl: 'registerapplicationform.component.html'
+    templateUrl: 'register-application-form.component.html'
 })
 export class EditApplicationFormComponent extends RegisterApplicationFormComponent implements OnInit {
     application: RegisteredApplication;
