+++ ui/src/main/webapp/src/app/components/migration-project-form.component.ts
@@ -2,11 +2,11 @@ import {Component, OnInit} from "@angular/core";
 import {ActivatedRoute, Router} from "@angular/router";
 
 import {MigrationProject} from "../windup-services";
-import {MigrationProjectService} from "../services/migrationproject.service";
-import {FormComponent} from "./formcomponent.component";
+import {MigrationProjectService} from "../services/migration-project.service";
+import {FormComponent} from "./form.component";
 
 @Component({
-    templateUrl: 'migrationprojectform.component.html'
+    templateUrl: 'migration-project-form.component.html'
 })
 export class MigrationProjectFormComponent extends FormComponent implements OnInit
 {
