+++ ui/src/main/webapp/src/app/components/progress-bar.component.ts
