+++ ui/src/main/webapp/src/app/components/register-application-form.component.html
