+++ ui/src/main/webapp/src/app/components/reports/technologies/technologies.report.ts
@@ -4,7 +4,7 @@ import {ActivatedRoute, Params}   from '@angular/router';
 import {TechReportService, StatsItem} from "./tech-report.service";
 
 import {ApplicationGroup} from "../../../../app/windup-services";
-import {ApplicationGroupService} from '../../../services/applicationgroup.service';
+import {ApplicationGroupService} from '../../../services/application-group.service';
 import {TechnologiesStatsModel} from '../../../generated/tsModels/TechnologiesStatsModel';
 
 @Component({
