+++ ui/src/main/webapp/src/app/services/analysis-context.service.ts
