+++ ui/src/main/webapp/src/app/services/application-group.service.ts
