+++ ui/src/main/webapp/src/app/services/registered-application.service.ts
