+++ ui/src/main/webapp/tests/app/registeredapplication.service.spec.ts
@@ -11,7 +11,7 @@ import 'rxjs/Rx';
 import {Constants} from '../../src/app/constants';
 
 import {KeyCloakServiceMock} from "./mocks/keycloak-service.mock";
-import {RegisteredApplicationService} from "../../src/app/services/registeredapplication.service";
+import {RegisteredApplicationService} from "../../src/app/services/registered-application.service";
 import {KeycloakService} from "../../src/app/services/keycloak.service";
 import {FileService} from "../../src/app/services/file.service";
 import {FileUploader, FileUploaderOptions} from "ng2-file-upload/ng2-file-upload";
